backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Banco de Dados
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Configurações
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Segurança / Auth
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9

# HTTP Client (OAuth)
httpx==0.27.0

# Redis / Cache
redis[asyncio]==5.0.4
celery[redis]==5.4.0

# AWS
boto3==1.34.101

# Firebase (Push Notifications)
firebase-admin==6.5.0

# Testes
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
